refactor: change commonj to ESM
