+++ package.json
@@ -11,7 +11,9 @@
     "read": "node src/fs/read.js",
     "rename": "node src/fs/rename.js",
     "env": "node src/cli/env.js",
-    "args": "node src/cli/args.js"
+    "args": "node src/cli/args.js",
+    "mjs": "node src/modules/cjsToEsm.mjs",
+
   },
   "repository": {
     "type": "git",
